this config works for now
-350ms watchdog
-write command every 10ms(!!!)

i think the way forward would be to turn on the sensor using GPIO and a transistor. still need to figure out the exact reason it perfectly functions after BOOTSEL but not in normal use. hint: it's not because of the watchdog reset. another hint: sleeping for like a second before resetting the UART fixes it.

so the problem is probably just that the sensor is bad and unreliable; and that nothing is bad with the code or the RP2040, and that by just waiting a few seconds the sensor doesn't send a bad data packet and can also respond to our transmissions easier.

the current solution effectively bruteforces the time the sensor needs to be "properly" initialized.

we should also figure out how long it takes for reliable data transmission to start, because the sensor likes to falsely claim that something is right in front of it as soon as data packets are enabled.